00-adk-intro/adk05a-mcp-instavibe/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.90.0
Flask==3.1.0
mcp[cli]==1.7.1
google-adk==0.4.0
python-dateutil==2.9.0.post0
